Move gogogate2 coordinator to separate module
