+++ homeassistant/components/gogogate2/common.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Awaitable, Callable, Mapping
+from collections.abc import Mapping
 from datetime import timedelta
 import logging
 from typing import Any, NamedTuple
@@ -24,16 +24,12 @@ from homeassistant.const import (
     CONF_USERNAME,
 )
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.debounce import Debouncer
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.httpx_client import get_async_client
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity, UpdateFailed
 
 from .const import DATA_UPDATE_COORDINATOR, DEVICE_TYPE_ISMARTGATE, DOMAIN, MANUFACTURER
+from .coordinator import DeviceDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -44,38 +40,6 @@ class StateData(NamedTuple):
     config_unique_id: str
     unique_id: str | None
     door: AbstractDoor | None
-
-
-class DeviceDataUpdateCoordinator(
-    DataUpdateCoordinator[GogoGate2InfoResponse | ISmartGateInfoResponse]
-):  # pylint: disable=hass-enforce-coordinator-module
-    """Manages polling for state changes from the device."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        logger: logging.Logger,
-        api: AbstractGateApi,
-        *,
-        name: str,
-        update_interval: timedelta,
-        update_method: Callable[
-            [], Awaitable[GogoGate2InfoResponse | ISmartGateInfoResponse]
-        ]
-        | None = None,
-        request_refresh_debouncer: Debouncer | None = None,
-    ) -> None:
-        """Initialize the data update coordinator."""
-        DataUpdateCoordinator.__init__(
-            self,
-            hass,
-            logger,
-            name=name,
-            update_interval=update_interval,
-            update_method=update_method,
-            request_refresh_debouncer=request_refresh_debouncer,
-        )
-        self.api = api
 
 
 class GoGoGate2Entity(CoordinatorEntity[DeviceDataUpdateCoordinator]):
