+++ homeassistant/components/gogogate2/cover.py
@@ -20,12 +20,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import (
-    DeviceDataUpdateCoordinator,
-    GoGoGate2Entity,
-    cover_unique_id,
-    get_data_update_coordinator,
-)
+from .common import GoGoGate2Entity, cover_unique_id, get_data_update_coordinator
+from .coordinator import DeviceDataUpdateCoordinator
 
 
 async def async_setup_entry(
