+++ homeassistant/components/gogogate2/sensor.py
@@ -16,12 +16,8 @@ from homeassistant.const import PERCENTAGE, EntityCategory, UnitOfTemperature
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import (
-    DeviceDataUpdateCoordinator,
-    GoGoGate2Entity,
-    get_data_update_coordinator,
-    sensor_unique_id,
-)
+from .common import GoGoGate2Entity, get_data_update_coordinator, sensor_unique_id
+from .coordinator import DeviceDataUpdateCoordinator
 
 SENSOR_ID_WIRED = "WIRE"
